Probability calculation

Calculate the probability of all potential response patterns for each respondent

p_thirt(n_item, n_block, n_person, dict, gamma, lambda, theta, psisq)

Arguments

n_item
number of items per block

n_block
number of blocks

n_person
number of respondents

dict
a data.frame with a dim column identifying traits by number 1, 2, 3... and an item column identifying item indicators for different traits by number 1, 2, 3...

gamma
a data.frame with one variable of length [total binary outcomes], row names should be of the format i-j for item pair ij

lambda
a data.frame with one variable of length [total items], row names should be of the format i for item i

theta
a data.frame of dimension [number of people X number of dimensions], row names should be of the format p for person p

psisq
a data.frame with one variable of length [total items], row names should be of the format i for item i

Value

a list of [permutation_list, probability], permutation_list is a matrix of dimension [number of permutations X number of items], probability is a matrix of dimension [person X permutation] of probability for each response pattern

Examples

if (FALSE) {
set.seed(202106)

#####################
## Test Conditions ##
#####################

n_item    = 4
n_block   = 1
n_dim     = 4
n_person  = 200
block_size <- n_item * (n_item - 1) / 2

################
## Parameters ##
################

# thresholds - gamma
# vector of length [total binary outcomes]
# row names should be of the format "i-j" for item pair ij
gamma   <- data.frame(gamma = round(runif(n = block_size * n_block, min = -1,  max = 1), digits = 3))
row.names(gamma) <- c("1-2", "1-3", "1-4", "2-3", "2-4", "3-4")

# loadings - lambda
# vector of length [total items]
# row names should be of the format "i" for item i
lambda  <- data.frame(lambda = round(runif(n = n_item * n_block, min = .65, max = .95), digits = 3))

# uniqueness - psi^2
# vector of length [total items]
# row names should be of the format "i" for item i
psisq   <- data.frame(psisq = round(1 - lambda^2, digits = 3))

# theta - trait scores
# matrix of dimension [number of people X number of dimensions]
# row names should be of the format "p" for person p
theta   <- as.data.frame(matrix(nrow = n_person, ncol = n_dim))
for (dim in seq(n_dim)) {
  theta[dim] <- round(rnorm(n = n_person, mean = 0, sd = 1), digits = 3)
} # END for dim LOOP

# create a dictionary with a "dim" column identifying traits
#   and an "item" column identifying item indicators for different traits
dict <- data.frame(
  dim = c(seq(n_dim)),
  item = c(seq(n_item * n_block)))

###################
## Test Function ##
###################

# test function with fixed parameters
prob <- p_thirt(n_item, n_block, n_person, dict,
                gamma, lambda, theta, psisq)

prob$permutation_list
summary(prob$probability)
}