Simulation tests

Examine parameter recovery and performance using simulated data

sim_test(
  n_person,
  n_item,
  n_neg,
  n_block,
  n_dim,
  n_iter,
  n_burnin,
  step_size_sd,
  tirt = FALSE,
  seed = 20220414
)

Arguments

n_person

integer indicating the number of people

n_item

integer or vector of length [n_block] indicating the number of items for each block

n_neg

integer or vector of length [n_block] indicating the number of negatively-keyed items for each block

n_block

integer indicating the number of blocks

n_dim

integer indicating the number of dimensions or traits

n_iter

integer indicating the number of MCMC iterations

n_burnin

integer indicating the number of MCMC burn-ins

step_size_sd

double indicating the step size of new parameter generation for random-walk Metropolis step

tirt

boolean indicating whether estimation results should be compared with the thurstonianIRT package

seed

random seed

Value

save a timestamped RData file in the current working directory

Examples

if (FALSE) {
sim_test(n_person = 10,
         n_item   = 4,
         n_neg    = 1,
         n_block  = 2,
         n_dim    = 4,
         n_iter   = 100,
         n_burnin = 1,
         step_size_sd = 0.1,
         tirt     = TRUE,
         seed     = 20220414)

}